fix: extract dialogs from buttons and use useDialog hook. Also remove vote information from CircularProgressWithLabel
