+++ packages/frontend/src/poker/components/dialogs/ResetVotesDialog.tsx
@@ -1,0 +1,47 @@
+import React from "react";
+import {
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogContentText,
+  DialogTitle,
+} from "@mui/material";
+import { DialogProps } from "../../../common/types/commonTypes";
+import { useFullscreen } from "../../../retro/hooks/useFullscreen";
+import { usePokerContext } from "../../context/PokerContext";
+
+export function ResetVotesDialog({ isOpen, close }: DialogProps) {
+  const { handleResetUserStory } = usePokerContext();
+  const fullScreen = useFullscreen();
+  function handleClick() {
+    handleResetUserStory();
+    close();
+  }
+
+  return (
+    <Dialog
+      fullWidth
+      maxWidth="sm"
+      fullScreen={fullScreen}
+      open={isOpen}
+      onClose={close}
+      aria-labelledby="poker-reset-dialog-title"
+    >
+      <DialogTitle id="poker-reset-dialog-title">Reset Votes</DialogTitle>
+      <DialogContent>
+        <DialogContentText>
+          You are about to reset all votes! Do you want to proceed?
+        </DialogContentText>
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={close} color="primary">
+          No
+        </Button>
+        <Button onClick={handleClick} color="error">
+          Yes
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
