+++ packages/frontend/src/retro/components/cards/EditCardButton.tsx
@@ -5,7 +5,7 @@ import { RetroCard } from "../../types/retroTypes";
 import { useIsPrivileged } from "../../hooks/useIsPrivileged";
 import { CardActionButton } from "./CardActionButton";
 import { EditCardDialog } from "../dialogs/EditCardDialog";
-import { useDialog } from "../../hooks/useDialog";
+import { useDialog } from "../../../common/hooks/useDialog";
 
 interface EditCardButtonProps extends IconButtonProps {
   card: RetroCard;
