+++ packages/frontend/src/retro/components/columns/column-header/CreateCardButton.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Add } from "@mui/icons-material";
 import { RetroColumn } from "../../../types/retroTypes";
 import { CreateCardDialog } from "../../dialogs/CreateCardDialog";
-import { useDialog } from "../../../hooks/useDialog";
 import { TooltipIconButton } from "../../../../common/TooltipIconButton";
+import { useDialog } from "../../../../common/hooks/useDialog";
 
 interface CreateCardButtonProps {
   column: RetroColumn;
