+++ packages/frontend/src/retro/components/columns/column-header/EditColumnMenuItem.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Edit } from "@mui/icons-material";
 import { ListItemIcon, ListItemText, MenuItem } from "@mui/material";
-import { useDialog } from "../../../hooks/useDialog";
 import { EditColumnDialog } from "../../dialogs/EditColumnDialog";
 import { RetroColumn } from "../../../types/retroTypes";
+import { useDialog } from "../../../../common/hooks/useDialog";
 
 interface EditColumnMenuItemProps {
   column: RetroColumn;
